Refactor input and display into modules
